+++ luxonis_ml/data/augmentations/__init__.py
@@ -1,6 +1,4 @@
-from .letterbox_resize import LetterboxResize
-from .mixup import MixUp
-from .mosaic import Mosaic4
+from .custom import LetterboxResize, MixUp, Mosaic4
 from .utils import Augmentations, TrainAugmentations, ValAugmentations
 
 __all__ = [
